.claude/worktrees/kind-hawking/requirements.txt
alpaca-py==0.28.0
pandas==2.2.2
python-dotenv==1.0.1
tzdata==2025.1
ta==0.11.0